requirements/base.txt
Django==5.2.6
django-environ==0.12.0
psycopg2==2.9.10
djangorestframework==3.16.1
drf-spectacular>=0.27.0
djangorestframework-simplejwt
celery==5.3.4
redis[hiredis]==5.0.1
django-redis
django-cors-headers
sib-api-v3-sdk
cloudinary
django-cloudinary-storage
pillow
requests